@@ -1,5 +1,5 @@
 class ApiController < ActionController::API
-
+  before_action :authenticate_user!
   rescue_from ActiveRecord::RecordNotFound, with: :not_found
 
   private
