@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 class ApplicationMailer < ActionMailer::Base
-  default from: 'from@example.com'
+  default from: email_address_with_name("platform@wildflowerschools.org", "Wildflower Platform")
   layout 'mailer'
 end
