@@ -14,7 +14,7 @@ module WildflowerPlatform
     config.load_defaults 7.0
 
     config.autoload_paths += %W(#{config.root}/lib)
-    config.eager_load_paths += %W(#{config.root}/lib)
+    # config.eager_load_paths += %W(#{config.root}/lib)
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
@@ -27,5 +27,7 @@ module WildflowerPlatform
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = true
+
+    config.action_mailer.deliver_later_queue_name = "mailers"
   end
 end
