@@ -2,7 +2,7 @@
 class CreateWorkflowDefinitionDependencies < ActiveRecord::Migration[7.0]
   def change
     # Dependencies are scope to a given workflow.
-    # can define for a process/step a prequisite process/step
+    # can define for a process/step a prerequisite process/step
     # am i asking can I be started?
     # or am I asking, what is next?
     # what is next is whatever can be started.
